refactor(isp-editor): extract DiffPreview and GoalTabPanel
